requirements.txt
Django==5.2.3
psycopg2-binary==2.9.10
djangorestframework==3.16.0
pytest==8.4.0
pytest-django==4.11.1
structlog==25.4.0
python-dotenv==1.1.0
django-extensions==4.1
Faker==37.4.0
requests==2.32.4
pytest-cov==6.2.1
